Share common ConfigureInfrastructure code for Azure Container Registry
