+++ src/Aspire.Hosting.Azure.ContainerRegistry/Aspire.Hosting.Azure.ContainerRegistry.csproj
@@ -9,6 +9,7 @@
 
   <ItemGroup>
     <Compile Include="$(RepoRoot)src\Shared\AzureRoleAssignmentUtils.cs" />
+    <Compile Include="$(RepoRoot)src\Shared\ContainerRegistryInfrastructure.cs" />
   </ItemGroup>
 
   <ItemGroup>
