+++ src/Aspire.Hosting.Foundry/Aspire.Hosting.Foundry.csproj
@@ -20,6 +20,7 @@
 
   <ItemGroup>
     <Compile Include="$(RepoRoot)src\Shared\AzureRoleAssignmentUtils.cs" />
+    <Compile Include="$(RepoRoot)src\Shared\ContainerRegistryInfrastructure.cs" />
     <Compile Remove="tools\**\*.cs" />
   </ItemGroup>
