+++ tests/Aspire.Hosting.Azure.Tests/Snapshots/AzureContainerAppsTests.DefaultContainerRegistryUsesAzdNamingWhenEnvironmentDoes.verified.bicep
@@ -17,3 +17,5 @@ resource env_acr 'Microsoft.ContainerRegistry/registries@2025-04-01' = {
 output name string = replace('acr-${resourceToken}', '-', '')
 
 output loginServer string = env_acr.properties.loginServer
+
+output id string = env_acr.id
