+++ tests/Aspire.Hosting.Azure.Tests/Snapshots/ExistingAzureResourceTests.SupportsExistingAzureContainerRegistryInRunMode.verified.bicep
@@ -7,7 +7,7 @@ resource acr 'Microsoft.ContainerRegistry/registries@2025-04-01' existing = {
   name: existingResourceName
 }
 
-output name string = acr.name
+output name string = existingResourceName
 
 output loginServer string = acr.properties.loginServer
